Fix filtering for bot changing name, and add a new one.
User-Agent: BlogPulse (ISSpider-3.0)

Access random pages; the related website (which cannot really be confirmed)
http://www.blogpulse.com/ shows a company making the usual marketing
analysis of blogs and stuff like that.

User-Agent: Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10.6; en-US; rv:1.9.2) Gecko/20100115 Firefox/3.6 (FlipboardBrowserProxy/0.0.5; +http://flipboard.com/browserproxy)

Tries to pass for a Firefox installation, but it's the crawler for the
Flipboad application for iPad; while the use seem a good one, it doesn't
respect Accept-Encoding, the linked URL is nowhere to be found (double 404
when accessing it), and the website has no email address to contact.