Add a second level of detection of bad robots, with regexp for those who can't be matched with @pm.
User-Agent: oBot

No information provided by the bot itself; the IP address it came from is
registered to Only Solutions in Germany; a no website listed on the whois
information either.

A quick search around produces this post:

http://www.webmasterworld.com/forum11/241.htm

and indicates http://www.onlysolutions.de/ as the home of the bot;
unfortunately the website itself is unavailable.

Missing any information about it, I have to take at face value the
declaration on the post above, and that is

> They scan the web looking for sites that infringe on copyrights and logos
> of their clients.