Add one more marketing bot: Ahrefs.
Requests look like this:

212.113.37.105 - - [01/Dec/2011:16:55:16 +0100] "GET /comments?article_id=4361 HTTP/1.1" 403 230 "-" "Mozilla/5.0 (compatible; AhrefsBot/1.0; +http://ahrefs.com/robot/)"

It probes _any_ link, even those forms are POST'd to, and does not do so
slowly, but one request after the other.

The company is based (standing to whois information) in Belize and provides
"backlinks analysis" for a fee, but the scanning IP comes from Ukraine.