Consider Anonymouse-bound requests to come from an open proxy.
The anonymouse service is one of many anonymisers, but a simple one at it
since it only fetches the requested page from one of its own IPs reduced to
the bare minimum of headers.

Instead of simply listing it on the bad robots list, make it automatically
an Open Proxy request so that it can't send POST requests.