Add semalt to the list of bad referrers.
Semalt is a "SEO" company that seems to crawl random websites'
homepages with a webdriver-like system (the request User-Agent is that
of Chrome, and it does not look incorrect. It also executes JavaScript
so it gets tracked by Analytics.

The problem is, beside not announcing themselves, that they use their
website as Referer for *all* the requests, causing it to jump up in
the list of inbound links for no good reason.

See:

http://www.hallaminternet.com/2014/remove-semalt-google-analytics/
http://en.forums.wordpress.com/topic/semaltcom-2?replies=43
http://wordpress.org/support/topic/how-to-block-semaltcom-from-visiting-your-wordpress-website