Stop ScreenerBot from harvesting pages.
Their website is pretty empty on what it should be, it is right now hosted
at Hetzner, but used to be hosted at Leaseweb last year. They only show the
list of IPs it's coming from on the website and provide a form of sorts to
block your website.

Threat as a malicious crawler.