Add rules to validate Chrome-Compression-Proxy.
The Data Saver Extension for Chrome is now available on desktop, and that
means more people are using it and it's easier for spammers to hide in
this.

Extract the valid forwarded IP address reported by the Proxy to check DNSBL
against.

Also if the request appears to be coming from the proxy, make sure the IP
addresses are actually valid. Similar rules can't exist for Opera (because
they don't mark themselves in any special way), but they can probably be
there for Amazon but I don't have a Kindle device to test.